Reorganized directories and namespaces.
