+++ Dietownik.ApplicationServices/API/Domain/Recipes/DeleteRecipeResponse.cs
@@ -1,0 +1,7 @@
+namespace Dietownik.ApplicationServices.API.Domain.Recipes
+{
+    public class DeleteRecipeResponse
+    {
+        
+    }
+}
