+++ Dietownik.ApplicationServices/API/Handlers/Products/DeleteProductHandler.cs
@@ -3,7 +3,7 @@ using System.Threading.Tasks;
 using AutoMapper;
 using Dietownik.ApplicationServices.API.Domain.Products;
 using Dietownik.DataAccess.CQRS;
-using Dietownik.DataAccess.CQRS.Commands;
+using Dietownik.DataAccess.CQRS.Commands.Products;
 using Dietownik.DataAccess.Entities;
 using MediatR;
