+++ Dietownik.ApplicationServices/API/Handlers/Recipes/AddRecipeIngredientHandler.cs
@@ -4,8 +4,8 @@ using AutoMapper;
 using Dietownik.ApplicationServices.API.Domain.Recipes;
 using Dietownik.DataAccess;
 using Dietownik.DataAccess.CQRS;
-using Dietownik.DataAccess.CQRS.Commands;
-using Dietownik.DataAccess.CQRS.Queries;
+using Dietownik.DataAccess.CQRS.Commands.Recipes;
+using Dietownik.DataAccess.CQRS.Queries.Products;
 using Dietownik.DataAccess.Entities;
 using MediatR;
