+++ Dietownik.ApplicationServices/API/Handlers/Recipes/DeleteRecipeHandler.cs
@@ -1,0 +1,7 @@
+namespace Dietownik.ApplicationServices.API.Handlers.Recipes
+{
+    public class DeleteRecipeHandler
+    {
+        
+    }
+}
