+++ Dietownik.DataAccess/CQRS/Commands/Products/DeleteProductCommand.cs
@@ -1,13 +1,13 @@
 using System.Threading.Tasks;
 using Dietownik.DataAccess.Entities;
 
-namespace Dietownik.DataAccess.CQRS.Commands
+namespace Dietownik.DataAccess.CQRS.Commands.Products
 {
-    public class UpdateProductCommand : CommandBase<Product, Product>
+    public class DeleteProductCommand : CommandBase<Product, Product>
     {
         public override async Task<Product> Execute(RecipeStorageContext context)
         {
-            context.Products.Update(this.Parameter);
+            context.Products.Remove(this.Parameter);
             await context.SaveChangesAsync();
             return this.Parameter;
         }
