+++ Dietownik.DataAccess/CQRS/Queries/Products/GetProductByIdQuery.cs
@@ -2,7 +2,7 @@ using System.Threading.Tasks;
 using Dietownik.DataAccess.Entities;
 using Microsoft.EntityFrameworkCore;
 
-namespace Dietownik.DataAccess.CQRS.Queries
+namespace Dietownik.DataAccess.CQRS.Queries.Products
 {
     public class GetProductByIdQuery : QueryBase<Product>
     {
